(feat) Initialize lib/ folder to contain socket configuration and action helpers
