+++ client/reducers/user.js
@@ -1,5 +1,7 @@
 import { Map } from 'immutable';
 
+import { handleUserInfo } from './userInfo';
+
 export default function (state = Map(), action) {
   // console.log('reducing user state:', state.toJS());
 
@@ -16,44 +18,6 @@ export default function (state = Map(), action) {
 
   return state;
 };
-
-export function handleUserInfo(state, action) {
-  switch (action.type) {
-    case 'REQUEST_USER_INFO':
-      return requestUserInfo(state, action);
-    case 'RECEIVE_USER_INFO':
-      return receiveUserInfo(state, action);
-    case 'REQUEST_USER_INFO_ERROR':
-      return requestUserInfoError(state, action);
-  }
-
-  return state;
-}
-
-function requestUserInfo(state) {
-  let updates = {
-    isFetchingUserInfo: true,
-  };
-
-  return state.merge(updates);
-}
-
-function receiveUserInfo(state, { entry }) {
-  let updates = {
-    ...entry,
-    isFetchingUserInfo: false,
-  };
-
-  return state.merge(updates);
-}
-
-function requestUserInfoError(state, { entry }) {
-  let updates = {
-    requestUserError: entry,
-  };
-
-  return state.merge(updates);
-}
 
 function setDriver(state, { entry }) {
   let updates = {
